Move test result and timer into modules
